@@ -1,9 +1,9 @@
 package com.tallogre.hanbaobao.Adapters;
 
 import android.os.Parcelable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentStatePagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentStatePagerAdapter;
 
 import com.tallogre.hanbaobao.Fragments.QuickDictionaryFragment;
 import com.tallogre.hanbaobao.Fragments.SettingsFragment;
@@ -33,7 +33,7 @@ public class TabViewPagerAdapter extends FragmentStatePagerAdapter {
     }
 
     @Override
-    public android.support.v4.app.Fragment getItem(int position) {
+    public Fragment getItem(int position) {
         return tabs.get(position);
     }
 
