@@ -4,28 +4,28 @@ import android.database.ContentObserver;
 import android.database.Cursor;
 import android.database.DataSetObserver;
 import android.os.Handler;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.widget.Filter;
 import android.widget.FilterQueryProvider;
 import android.widget.Filterable;
 
 /**
- * Provide a {@link android.support.v7.widget.RecyclerView.Adapter} implementation with cursor
+ * Provide a {@link RecyclerView.Adapter} implementation with cursor
  * support.
  * <p/>
  * Child classes only need to implement {@link #onCreateViewHolder(android.view.ViewGroup, int)} and
- * {@link #onBindViewHolderCursor(android.support.v7.widget.RecyclerView.ViewHolder, android.database.Cursor)}.
+ * {@link #onBindViewHolderCursor(RecyclerView.ViewHolder, android.database.Cursor)}.
  * <p/>
  * This class does not implement deprecated fields and methods from CursorAdapter! Incidentally,
  * only {@link android.widget.CursorAdapter#FLAG_REGISTER_CONTENT_OBSERVER} is available, so the
  * flag is implied, and only the Adapter behavior using this flag has been ported.
  *
  * @param <VH> {@inheritDoc}
- * @see android.support.v7.widget.RecyclerView.Adapter
+ * @see RecyclerView.Adapter
  * @see android.widget.CursorAdapter
  * @see android.widget.Filterable
  */
-public abstract class CursorRecyclerAdapter<VH extends android.support.v7.widget.RecyclerView.ViewHolder> extends RecyclerView.Adapter<VH>
+public abstract class CursorRecyclerAdapter<VH extends RecyclerView.ViewHolder> extends RecyclerView.Adapter<VH>
         implements Filterable, CursorFilter.CursorFilterClient {
     private static final String ROW_ID_COLUMN_NAME = "rowid";
     private boolean mDataValid;
